+++ packages/design-system-mui/components/ButtonWithLoading/index.tsx
@@ -1,22 +1,20 @@
+import { FC } from 'react'
+
 import { Button } from '@mui/material'
 
 import { CircularProgress } from './styled'
 import { IButtonWitthLoadingProps } from './types'
 
-function ButtonWithLoading({
-  loading = false,
+const ButtonWithLoading: FC<IButtonWitthLoadingProps> = ({
   children,
-  loadingChildren = <CircularProgress size="20px" />,
-  form,
+  isLoading = false,
+  loadingComponent = <CircularProgress size="20px" />,
   ...props
-}: IButtonWitthLoadingProps) {
-  const isLoading = loading || form?.formState.isSubmitting
-  return (
-    <Button disabled={isLoading} {...props}>
-      {children}
-      {isLoading && loadingChildren}
-    </Button>
-  )
-}
+}) => (
+  <Button disabled={isLoading} {...props}>
+    {children}
+    {isLoading && loadingComponent}
+  </Button>
+)
 
 export default ButtonWithLoading